refact(kt-comment): move CommentActionsReply and CommentActionsOptions into CommentActions
